Master v6

+++ simd_math_tests/generate_plots.sh
@@ -5,14 +5,18 @@ do
     cd $dir
     
     echo $(pwd)
-    rm -rf build
-    mkdir -p build
-    cmake -S . -B ./build -GNinja -DCMAKE_BUILD_TYPE=Release > /dev/null
-    cd build
-    ninja
-    cd ..
-    build/exe
-    rm -rf build *.log
+    if [ -d "plots" ]; then
+        echo "plots for $dir already exists"
+    else
+        rm -rf build
+        mkdir -p build
+        cmake -S . -B ./build -GNinja -DCMAKE_BUILD_TYPE=Release > /dev/null
+        cd build
+        ninja
+        cd ..
+        build/exe $SIMD_END > /dev/null
+        rm -rf build *.log
+    fi
     cd ..
 done
 python3 generate_plots.py
